Move instructions to Nextstrain docs page
This repository now only contains files necessary to run the tutorial
steps, similar to the zika-tutorial repository.

I've opted to keep the environment setup steps in the README since it's
more related to the repo than the tutorial steps. The options were moved
into their own 3rd-level headings to allow the GitHub markdown alerts to
render, as they cannot be nested within other elements.¹

The link to the doc page is a docs PR preview build for now. I'll update
the link once the PR is merged.

¹ <https://docs.github.com/en/get-started/writing-on-github/getting-started-with-writing-and-formatting-on-github/basic-writing-and-formatting-syntax#alerts>