Restore GitPod option
Despite the GitPod flex vs. classic comparison page¹ making it sound
like the hosted plan is not available to new users, it actually still
is.

Other UF workshop modules will be using GitPod, so add it as an option
for attendees to use the same environment as they used previously.

¹ <https://www.gitpod.io/flex-or-classic>